Refactor maven project structure
- create itest module
- rename '-main' to '-parent'
- make child module poms depend on parent
- remove itest dependencies from addon pom
- dependencymanagement, common properties, scm, license in parent pom
- upgrade vaadin, deltaspike, cdi, servlet, etc.. versions
- drop zip assembly for vaadin directory
- remove checkstyle
