--- vaadin-cdi/assembly/assembly.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<assembly
-	xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-	<id>addon</id>
-
-	<!-- Note: the base directory for this assembly descriptor is one of the 
-		license specific subprojects -->
-
-	<formats>
-		<format>zip</format>
-	</formats>
-
-	<!-- Do not use because we must put META-INF/MANIFEST.MF there. -->
-	<includeBaseDirectory>false</includeBaseDirectory>
-
-	<fileSets>
-		<fileSet>
-			<directory>.</directory>
-			<includes>
-				<include>license-*.txt</include>
-				<include>licensing.txt</include>
-			</includes>
-		</fileSet>
-		<fileSet>
-			<directory>target</directory>
-			<outputDirectory></outputDirectory>
-			<includes>
-				<include>*.jar</include>
-                <include>*.pdf</include>
-			</includes>
-		</fileSet>
-	</fileSets>
-
-	<files>
-		<!-- This is vaadin.com/directory related manifest needed in the zip package -->
-		<file>
-			<source>assembly/MANIFEST.MF</source>
-			<outputDirectory>META-INF</outputDirectory>
-			<filtered>true</filtered>
-		</file>
-	</files>
-</assembly>
